Add files via upload
Tic Tac Toe is a classic paper-and-pencil game played between two players, traditionally using a 3x3 grid. The objective is to be the first to form a horizontal, vertical, or diagonal line of your own symbols - typically "X" and "O" - within the grid. This repository contains a simple implementation of the classic Tic Tac Toe game built in [programming language of choice]. The game provides a console-based interface for players to interact with. #yenepoyauniversity #Tictactoe #gamedevelpoment